+++ packages/api-contract/src/bundle.ts
@@ -1,0 +1,7 @@
+// Copyright 2017-2021 @polkadot/api-contract authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { Abi } from './Abi';
+export { packageInfo } from './packageInfo';
+export { Blueprint as BlueprintPromise, Code as CodePromise, Contract as ContractPromise } from './promise';
+export { Blueprint as BlueprintRx, Code as CodeRx, Contract as ContractRx } from './rx';
