+++ packages/api-contract/src/index.ts
@@ -3,7 +3,4 @@
 
 import './detectPackage';
 
-export { Abi } from './Abi';
-export { packageInfo } from './packageInfo';
-export { Blueprint as BlueprintPromise, Code as CodePromise, Contract as ContractPromise } from './promise';
-export { Blueprint as BlueprintRx, Code as CodeRx, Contract as ContractRx } from './rx';
+export * from './bundle';
