+++ packages/api/src/bundle.ts
@@ -1,0 +1,10 @@
+// Copyright 2017-2021 @polkadot/api authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { Keyring } from '@polkadot/keyring';
+export { WsProvider } from '@polkadot/rpc-provider/ws';
+
+export { packageInfo } from './packageInfo';
+export { ApiPromise } from './promise';
+export { SubmittableResult } from './submittable';
+export { ApiRx } from './rx';
