+++ packages/api/src/index.ts
@@ -3,10 +3,4 @@
 
 import './detectPackage';
 
-export { Keyring } from '@polkadot/keyring';
-export { WsProvider } from '@polkadot/rpc-provider/ws';
-
-export { packageInfo } from './packageInfo';
-export { ApiPromise } from './promise';
-export { SubmittableResult } from './submittable';
-export { ApiRx } from './rx';
+export * from './bundle';
