+++ packages/metadata/src/bundle.ts
@@ -1,0 +1,6 @@
+// Copyright 2017-2021 @polkadot/metadata authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { Metadata } from './Metadata';
+export { decorateConstants, decorateExtrinsics, decorateStorage, expandMetadata } from './decorate';
+export { packageInfo } from './packageInfo';
