+++ packages/metadata/src/index.ts
@@ -3,6 +3,4 @@
 
 import './detectPackage';
 
-export { Metadata } from './Metadata';
-export { decorateConstants, decorateExtrinsics, decorateStorage, expandMetadata } from './decorate';
-export { packageInfo } from './packageInfo';
+export * from './bundle';
