+++ packages/rpc-provider/src/bundle.ts
@@ -1,0 +1,6 @@
+// Copyright 2017-2021 @polkadot/rpc-provider authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export { HttpProvider } from './http';
+export { packageInfo } from './packageInfo';
+export { WsProvider } from './ws';
