+++ packages/rpc-provider/src/index.ts
@@ -3,6 +3,4 @@
 
 import './detectPackage';
 
-export { HttpProvider } from './http';
-export { packageInfo } from './packageInfo';
-export { WsProvider } from './ws';
+export * from './bundle';
