+++ packages/typegen/src/bundle.ts
@@ -1,6 +1,5 @@
 // Copyright 2017-2021 @polkadot/typegen authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import './detectPackage';
-
-export * from './bundle';
+export { packageInfo } from './packageInfo';
+export { formatType } from './util/formatting';
