+++ packages/types/src/bundle.ts
@@ -1,0 +1,7 @@
+// Copyright 2017-2021 @polkadot/types authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './codec';
+export * from './create';
+export * from './index.types';
+export { packageInfo } from './packageInfo';
