+++ packages/types/src/index.ts
@@ -3,7 +3,4 @@
 
 import './detectPackage';
 
-export * from './codec';
-export * from './create';
-export * from './index.types';
-export { packageInfo } from './packageInfo';
+export * from './bundle';
